fix(build): move getBackupTimestamp to shared-with-frontend and update imports  #7320
