+++ electron/shared-with-frontend/get-backup-timestamp.spec.ts
