+++ electron/shared-with-frontend/get-backup-timestamp.ts
